+++ lib/collection/src/operations/universal_query/collection_query.rs
@@ -781,322 +781,3 @@ mod from_rest {
         }
     }
 }
-
-pub mod from_grpc {
-    use api::grpc::conversions::json_path_from_proto;
-    use api::grpc::qdrant::{self as grpc};
-    use tonic::Status;
-
-    use super::*;
-
-    impl TryFrom<api::grpc::qdrant::QueryPointGroups> for CollectionQueryGroupsRequest {
-        type Error = Status;
-
-        fn try_from(value: api::grpc::qdrant::QueryPointGroups) -> Result<Self, Self::Error> {
-            let grpc::QueryPointGroups {
-                collection_name: _,
-                prefetch,
-                query,
-                using,
-                filter,
-                params,
-                score_threshold,
-                with_payload,
-                with_vectors,
-                lookup_from,
-                limit,
-                group_size,
-                group_by,
-                with_lookup,
-                read_consistency: _,
-                timeout: _,
-                shard_key_selector: _,
-            } = value;
-
-            let request = CollectionQueryGroupsRequest {
-                prefetch: prefetch
-                    .into_iter()
-                    .map(TryFrom::try_from)
-                    .collect::<Result<_, _>>()?,
-                query: query.map(TryFrom::try_from).transpose()?,
-                using: using.unwrap_or(DEFAULT_VECTOR_NAME.to_string()),
-                filter: filter.map(TryFrom::try_from).transpose()?,
-                score_threshold,
-                with_vector: with_vectors
-                    .map(From::from)
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_WITH_VECTOR),
-                with_payload: with_payload
-                    .map(TryFrom::try_from)
-                    .transpose()?
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_WITH_PAYLOAD),
-                lookup_from: lookup_from.map(From::from),
-                group_by: json_path_from_proto(&group_by)?,
-                group_size: group_size
-                    .map(|s| s as usize)
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_GROUP_SIZE),
-                limit: limit
-                    .map(|l| l as usize)
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_LIMIT),
-                params: params.map(From::from),
-                with_lookup: with_lookup.map(TryFrom::try_from).transpose()?,
-            };
-            Ok(request)
-        }
-    }
-
-    impl TryFrom<api::grpc::qdrant::QueryPoints> for CollectionQueryRequest {
-        type Error = Status;
-
-        fn try_from(value: api::grpc::qdrant::QueryPoints) -> Result<Self, Self::Error> {
-            let grpc::QueryPoints {
-                collection_name: _,
-                prefetch,
-                query,
-                using,
-                filter,
-                params,
-                score_threshold,
-                limit,
-                offset,
-                with_payload,
-                with_vectors,
-                read_consistency: _,
-                shard_key_selector: _,
-                lookup_from,
-                timeout: _,
-            } = value;
-
-            let request = CollectionQueryRequest {
-                prefetch: prefetch
-                    .into_iter()
-                    .map(TryFrom::try_from)
-                    .collect::<Result<_, _>>()?,
-                query: query.map(TryFrom::try_from).transpose()?,
-                using: using.unwrap_or(DEFAULT_VECTOR_NAME.to_string()),
-                filter: filter.map(TryFrom::try_from).transpose()?,
-                score_threshold,
-                limit: limit
-                    .map(|l| l as usize)
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_LIMIT),
-                offset: offset
-                    .map(|o| o as usize)
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_OFFSET),
-                params: params.map(From::from),
-                with_vector: with_vectors
-                    .map(From::from)
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_WITH_VECTOR),
-                with_payload: with_payload
-                    .map(TryFrom::try_from)
-                    .transpose()?
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_WITH_PAYLOAD),
-                lookup_from: lookup_from.map(From::from),
-            };
-            Ok(request)
-        }
-    }
-
-    impl TryFrom<grpc::PrefetchQuery> for CollectionPrefetch {
-        type Error = Status;
-
-        fn try_from(value: grpc::PrefetchQuery) -> Result<Self, Self::Error> {
-            let grpc::PrefetchQuery {
-                prefetch,
-                query,
-                using,
-                filter,
-                params,
-                score_threshold,
-                limit,
-                lookup_from,
-            } = value;
-
-            let collection_query = Self {
-                prefetch: prefetch
-                    .into_iter()
-                    .map(TryFrom::try_from)
-                    .collect::<Result<_, _>>()?,
-                query: query.map(TryFrom::try_from).transpose()?,
-                using: using.unwrap_or(DEFAULT_VECTOR_NAME.to_string()),
-                filter: filter.map(TryFrom::try_from).transpose()?,
-                score_threshold,
-                limit: limit
-                    .map(|l| l as usize)
-                    .unwrap_or(CollectionQueryRequest::DEFAULT_LIMIT),
-                params: params.map(From::from),
-                lookup_from: lookup_from.map(From::from),
-            };
-
-            Ok(collection_query)
-        }
-    }
-
-    impl TryFrom<grpc::Query> for Query {
-        type Error = Status;
-
-        fn try_from(value: grpc::Query) -> Result<Self, Self::Error> {
-            use api::grpc::qdrant::query::Variant;
-
-            let variant = value
-                .variant
-                .ok_or_else(|| Status::invalid_argument("Query variant is missing"))?;
-
-            let query = match variant {
-                Variant::Nearest(nearest) => {
-                    Query::Vector(VectorQuery::Nearest(TryFrom::try_from(nearest)?))
-                }
-                Variant::Recommend(recommend) => Query::Vector(TryFrom::try_from(recommend)?),
-                Variant::Discover(discover) => Query::Vector(TryFrom::try_from(discover)?),
-                Variant::Context(context) => Query::Vector(TryFrom::try_from(context)?),
-                Variant::OrderBy(order_by) => Query::OrderBy(OrderBy::try_from(order_by)?),
-                Variant::Fusion(fusion) => Query::Fusion(Fusion::try_from(fusion)?),
-                Variant::Sample(sample) => Query::Sample(Sample::try_from(sample)?),
-            };
-
-            Ok(query)
-        }
-    }
-
-    impl TryFrom<grpc::RecommendInput> for VectorQuery<VectorInputInternal> {
-        type Error = Status;
-
-        fn try_from(value: grpc::RecommendInput) -> Result<Self, Self::Error> {
-            let grpc::RecommendInput {
-                positive,
-                negative,
-                strategy,
-            } = value;
-
-            let positives = positive
-                .into_iter()
-                .map(TryFrom::try_from)
-                .collect::<Result<Vec<_>, _>>()?;
-            let negatives = negative
-                .into_iter()
-                .map(TryFrom::try_from)
-                .collect::<Result<Vec<_>, _>>()?;
-
-            let reco_query = RecoQuery::new(positives, negatives);
-
-            let strategy = strategy
-                .and_then(|x|
-                    // XXX: Invalid values silently converted to None
-                    grpc::RecommendStrategy::try_from(x).ok())
-                .map(RecommendStrategy::from)
-                .unwrap_or_default();
-
-            let query = match strategy {
-                RecommendStrategy::AverageVector => VectorQuery::RecommendAverageVector(reco_query),
-                RecommendStrategy::BestScore => VectorQuery::RecommendBestScore(reco_query),
-            };
-
-            Ok(query)
-        }
-    }
-
-    impl TryFrom<grpc::DiscoverInput> for VectorQuery<VectorInputInternal> {
-        type Error = Status;
-
-        fn try_from(value: grpc::DiscoverInput) -> Result<Self, Self::Error> {
-            let grpc::DiscoverInput { target, context } = value;
-
-            let target = VectorInputInternal::try_from(
-                target
-                    .ok_or_else(|| Status::invalid_argument("DiscoverInput target is missing"))?,
-            )?;
-
-            let grpc::ContextInput { pairs } = context
-                .ok_or_else(|| Status::invalid_argument("DiscoverInput context is missing"))?;
-
-            let context = pairs
-                .into_iter()
-                .map(context_pair_from_grpc)
-                .collect::<Result<_, _>>()?;
-
-            Ok(VectorQuery::Discover(DiscoveryQuery::new(target, context)))
-        }
-    }
-
-    impl TryFrom<grpc::ContextInput> for VectorQuery<VectorInputInternal> {
-        type Error = Status;
-
-        fn try_from(value: grpc::ContextInput) -> Result<Self, Self::Error> {
-            let context_query = context_query_from_grpc(value)?;
-
-            Ok(VectorQuery::Context(context_query))
-        }
-    }
-
-    impl TryFrom<grpc::VectorInput> for VectorInputInternal {
-        type Error = Status;
-
-        fn try_from(value: grpc::VectorInput) -> Result<Self, Self::Error> {
-            use api::grpc::qdrant::vector_input::Variant;
-
-            let variant = value
-                .variant
-                .ok_or_else(|| Status::invalid_argument("VectorInput variant is missing"))?;
-
-            let vector_input = match variant {
-                Variant::Id(id) => VectorInputInternal::Id(TryFrom::try_from(id)?),
-                Variant::Dense(dense) => {
-                    VectorInputInternal::Vector(VectorInternal::Dense(From::from(dense)))
-                }
-                Variant::Sparse(sparse) => {
-                    VectorInputInternal::Vector(VectorInternal::Sparse(From::from(sparse)))
-                }
-                Variant::MultiDense(multi_dense) => VectorInputInternal::Vector(
-                    // TODO(universal-query): Validate at API level
-                    VectorInternal::MultiDense(From::from(multi_dense)),
-                ),
-                Variant::Document(_) => {
-                    return Err(Status::invalid_argument(
-                        "Document inference is not implemented",
-                    ))
-                }
-                Variant::Image(_) => {
-                    return Err(Status::invalid_argument(
-                        "Image inference is not implemented",
-                    ))
-                }
-                Variant::Object(_) => {
-                    return Err(Status::invalid_argument(
-                        "Object inference is not implemented",
-                    ))
-                }
-            };
-
-            Ok(vector_input)
-        }
-    }
-
-    /// Circular dependencies prevents us from implementing `TryFrom` directly
-    fn context_query_from_grpc(
-        value: grpc::ContextInput,
-    ) -> Result<ContextQuery<VectorInputInternal>, Status> {
-        let grpc::ContextInput { pairs } = value;
-
-        Ok(ContextQuery {
-            pairs: pairs
-                .into_iter()
-                .map(context_pair_from_grpc)
-                .collect::<Result<_, _>>()?,
-        })
-    }
-
-    /// Circular dependencies prevents us from implementing `TryFrom` directly
-    fn context_pair_from_grpc(
-        value: grpc::ContextInputPair,
-    ) -> Result<ContextPair<VectorInputInternal>, Status> {
-        let grpc::ContextInputPair { positive, negative } = value;
-
-        let positive =
-            positive.ok_or_else(|| Status::invalid_argument("ContextPair positive is missing"))?;
-        let negative =
-            negative.ok_or_else(|| Status::invalid_argument("ContextPair negative is missing"))?;
-
-        Ok(ContextPair {
-            positive: VectorInputInternal::try_from(positive)?,
-            negative: VectorInputInternal::try_from(negative)?,
-        })
-    }
-}
