+++ src/common/inference/mod.rs
@@ -1,1 +1,3 @@
-pub mod conversion;
+pub mod update_requests;
+pub mod query_requests_rest;
+pub mod query_requests_grpc;
